[osprey.compiler] Add support for collection and comprehension initializers
A collection initializer is simialr to an object initializer, but with square brackets instead of curly:

    new MyCollection with [1, 2, 3]

As the name suggests, it's used to initialize collections. Each element in the initializer becomes an argument to the constructed type's `add` method. More than one argument can be specified with parentheses:

    new MyHash with [(k1, v1), (k2, v2)]

A comprehension initializer is essentially the same, plus a list comprehension. Thus you can do this:

    new MyCollection with [v for v in values]

and indeed this:

    new MyHash with [
      (k, value),
      (k.toUpper(), value)
      for k in keys
      where k != null
    ]

Note how the element initializers still require parentheses when multiple arguments are involved, and that it is possible to add multiple entries in the same comprehension initializer.