Remove operators $ and #
These were initially intended to be reserved as programmer-defined operators, with no predetermined meanings. However, this naturally makes the operators meaningless. Anyone who reads code that uses these operators would have to know what they mean in the particular domain, which is not likely to be obvious. Named methods are a much better alternative to meaningless punctuation.