Remove wrapped tree entirely
The wrapped tree was added without much thought put into it at all. The idea at first was simply to have somewhere to put stateful information, but the problem with the wrapped tree was that it had to contain *everything* the regular tree has, and more. The wrapped tree will instead be replaced by a simplified intermediate tree, which will make code analysis and compilation much easier.