SYNTAX CHANGE: Use \u{..} for Unicode escapes
The new syntax is inherently delimited on both sides, making it much easier to read (even with syntax highlighting), and accepts any number of hexadecimal digits. The eight digits in \Uxxxxxxxx escapes were too numerous to be comfortable, especially as the highest code point is U+10FFFF, and the four in \uxxxx are not enough for non-BMP characters.