Remove ParseNode.parent and related functionality
This property (and related functionality, e.g. claim(), findNearest())
was added because it was felt to be sorely needed in the Osprey compiler
in C#. However, not much thought was actually given to how it would be
used. It had several problems:

* Setting the `parent` (of every single node the parser constructs)
  involves, among other things, enumerating every single NodeList and
  lots of unnecessary null checks. Performance suffered, albeit only to
  a tiny extent.
* The parse tree purports to be immutable, but the constructor of every
  node type can potentially alter the tree. One could render the tree
  unstable by simply passing a node into some random node constructor
  (maliciously or unintentionally), perhaps while hoping to reuse an
  Annotation or similar.
* Since NodeList is not a ParseNode, all NodeLists were effectively
  hidden from the tree when walking up. This could lead to surprising
  situations.

Everything that `parent` could accomplish can be expressed more clearly
and flexiby inside a ParseTreeVisitor/Walker, using techniques such as
context stacks or the `arg` argument that all visitors accept.