Remove unsafe constructors from ParameterList and ArgumentList
Allowing anyone to specify properties of parameter and argument lists
that are derived from the parameters/arguments makes it possible to
create inconsistent lists. The backing fields are made internal so the
parser can still set them. For parse nodes generated by user code, the
performance impact of calculating these properties (on demand) is deemed
negligible.